requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
chromadb>=0.4.22
python-dotenv>=1.0.0
pyyaml>=6.0
loguru>=0.7.0
